+++ Graphics/Image.cpp
@@ -482,7 +482,7 @@ void Image::initialize(Graphics* graphics, std::string textureName)
 	spriteData->texture = graphics->getTexture(textureName);
 	this->cols = 1;
 	spriteData->rect->setLeft(0, 0, (float) spriteData->texture->image->w, (float) spriteData->texture->image->h);
-	this->rec.set(Vector3D(spriteData->texture->image->w*0.5f, spriteData->texture->image->h*0.5f), Vector3D(spriteData->texture->image->w, spriteData->texture->image->h), 0);
+	this->rec.set(Vector3D(spriteData->texture->image->w*0.5f, spriteData->texture->image->h*0.5f), Vector3D((float) spriteData->texture->image->w, (float) spriteData->texture->image->h), 0);
 	spriteData->textureName = textureName;
 	isInitialized = true;
 }
@@ -493,7 +493,7 @@ void Image::initialize(Graphics* graphics, mySurface* surface, std::string text)
 	spriteData->texture = surface;
 	this->cols = 1;
 	spriteData->rect->setLeft(0, 0, (float) spriteData->texture->image->w, (float) spriteData->texture->image->h);
-	this->rec.set(Vector3D(spriteData->texture->image->w*0.5f, spriteData->texture->image->h*0.5f), Vector3D(spriteData->texture->image->w, spriteData->texture->image->h), 0);
+	this->rec.set(Vector3D(spriteData->texture->image->w*0.5f, spriteData->texture->image->h*0.5f), Vector3D((float) spriteData->texture->image->w, (float) spriteData->texture->image->h), 0);
 	spriteData->textureName = text;
 	spriteData->isText = true;
 	isInitialized = true;
